made changes based on pr code review comments
