+++ AngularApp/projects/applens/src/app/modules/dashboard/onboarding-flow/onboarding-flow.component.html
@@ -183,7 +183,7 @@
         <update-detector-references [id] = "id" [Branch] = "Branch" [resourceId] = "resourceId" [mode] = "mode" 
         [dataSource] = "dataSource"  [timeRange] = "timeRange"  [compilationPackage]  = "compilationPackage"
         [queryResponse] = "queryResponse"  [userName] = "userName" [useAutoMergeText] = "useAutoMergeText"
-        [PRLink] = "PRLink"  [defaultBranch] = "defaultBranch" [owners] = "owners"  [DevopsConfig] = "DevopsConfig"
+        [defaultBranch] = "defaultBranch" [owners] = "owners"  [DevopsConfig] = "DevopsConfig"
         [detectorReferencesDialogHidden]  = "detectorReferencesDialogHidden" 
         [detectorReferencesList] = "detectorReferencesList"
         [gistCommitVersion] = "gistCommitVersion"></update-detector-references>
